Add functions to randomize and select colorButtons.
Reorganized how the colorButton constructor initializes. Colors now start off as null, but are then set using a new scramble() method.
This method can be used to select new random colors for a colorButton repeatedly.
Added a randomBox(array) function to select a random item from an array and returns the item selected.
Added a .color method to targetNum, similar to the ones in colorButton, it returns the current .red, .green, and .blue values in a "rgb(#,#,#)" format.