Improved the random color formula
Previously, the scramble function simply chose random values for red, green, and blue.
Scramble will now check whether the numbers generated are too close to the target number,
if they are, it will "reroll" until it creates an acceptable value (currently limited to 25 tries).
In easy mode, the values must be more than 50 points away from the target's value.
In normal mode, the values must be more than 25 points.
This can be controlled using the 'threshold' property in the difficulty object.